# api/openapi.yaml
components:
  schemas:
...
    DomainSearchHTTPResponse:
      additionalProperties: true
      properties:
        cache_revision:
          type: string
        candidate_snapshot_id:
          type: string
        next_cursor:
          type: string
        rows:
          items:
            $ref: "#/components/schemas/DomainSummary"
          type:
            - array
            - "null"
        search_provenance:
          $ref: "#/components/schemas/SearchProvenance"
        total_count:
          format: int64
          type: integer
      required:
        - rows
        - total_count
        - cache_revision
        - search_provenance
      type: object
    DomainSummary:
      additionalProperties: true
      properties:
        activity_count:
          format: int64
          type: integer
        cache_revision:
          type: string
        domain:
          type: string
        file_count:
          format: int64
          type: integer
        first_at:
          format: date-time
          type: string
        last_at:
          format: date-time
          type: string
        person_count:
          format: int64
          type: integer
        source_counts:
          items:
            $ref: "#/components/schemas/SourceCount"
          type:
            - array
            - "null"
      required:
        - domain
        - activity_count
        - person_count
        - file_count
        - source_counts
        - first_at
        - last_at
        - cache_revision
      type: object
...
    ErrorResponse:
      additionalProperties: true
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
      type: object
...
    ExploreCacheUnavailableResponse:
      additionalProperties: true
      properties:
        error:
          type: string
        message:
          type: string
        readiness:
          enum:
            - absent
            - interrupted
            - stale_schema
            - drifted
          type: string
        recovery_action:
          type: string
      required:
        - error
        - message
        - readiness
        - recovery_action
      type: object
...
    FileSearchHTTPRequest:
      additionalProperties: false
      properties:
        cursor:
          type: string
        filename_query:
          type: string
        limit:
          format: int64
          maximum: 500
          minimum: 0
          type: integer
        mime_families:
          items:
            type: string
          type:
            - array
            - "null"
        predicate:
          $ref: "#/components/schemas/ExploreHTTPRequest"
        sort:
          $ref: "#/components/schemas/FileSearchSort"
      required:
        - predicate
        - sort
      type: object
...
    IdentitySearchHTTPRequest:
      additionalProperties: false
      properties:
        cursor:
          type: string
        identity_query:
          type: string
        limit:
          format: int64
          maximum: 500
          minimum: 0
          type: integer
        predicate:
          $ref: "#/components/schemas/ExploreHTTPRequest"
        sort:
          $ref: "#/components/schemas/IdentitySearchSort"
      required:
        - predicate
        - sort
      type: object
...
  securitySchemes:
    apiKey:
      in: header
      name: X-Api-Key
      type: apiKey
info:
  title: msgvault API
  version: 1.24.0
openapi: 3.1.0
paths:
...
  /api/v1/domains/search:
    post:
      operationId: searchDomains
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IdentitySearchHTTPRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DomainSearchHTTPResponse"
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Error
        "409":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Error
        "503":
          content:
            application/json:
              schema:
                anyOf:
                  - $ref: "#/components/schemas/ExploreCacheUnavailableResponse"
                  - $ref: "#/components/schemas/ErrorResponse"
          description: Service Unavailable
        default:
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Error
      security:
        - apiKey: []
      summary: Search analytical domains
      tags:
        - Exploration
  /api/v1/domains/{domain}:
    get:
      operationId: getDomain
      parameters:
        - description: Exact normalized domain fact
          in: path
          name: domain
          required: true
          schema:
            type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DomainSummary"
          description: OK
        default:
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Error
      security:
        - apiKey: []
      summary: Get one analytical domain
      tags:
        - API
  /api/v1/domains/{domain}/files/search:
    post:
      operationId: searchDomainFiles
      parameters:
        - description: Exact normalized domain fact
          in: path
          name: domain
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FileSearchHTTPRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileSearchHTTPResponse"
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Error
        "409":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Error
        "503":
          content:
            application/json:
              schema:
                anyOf:
                  - $ref: "#/components/schemas/ExploreCacheUnavailableResponse"
                  - $ref: "#/components/schemas/ErrorResponse"
          description: Service Unavailable
        default:
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Error
      security:
        - apiKey: []
      summary: Search one domain's analytical files
      tags:
        - Exploration
  /api/v1/domains/{domain}/summary:
    post:
      operationId: getDomainContextSummary
      parameters:
        - description: Exact normalized domain fact
          in: path
          name: domain
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExploreHTTPRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DomainContextSummaryHTTPResponse"
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Error
        "409":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Error
        "503":
          content:
            application/json:
              schema:
                anyOf:
                  - $ref: "#/components/schemas/ExploreCacheUnavailableResponse"
                  - $ref: "#/components/schemas/ErrorResponse"
          description: Service Unavailable
        default:
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Error
      security:
        - apiKey: []
      summary: Get one domain's contextual analytical summary
      tags:
        - Exploration
  /api/v1/domains/{domain}/timeline:
    post:
      operationId: getDomainTimeline
      parameters:
        - description: Exact normalized domain fact
          in: path
          name: domain
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExploreHTTPRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExploreHTTPResponse"
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Error
        "409":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Error
        "503":
          content:
            application/json:
              schema:
                anyOf:
                  - $ref: "#/components/schemas/ExploreCacheUnavailableResponse"
                  - $ref: "#/components/schemas/ErrorResponse"
          description: Service Unavailable
        default:
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Error
      security:
        - apiKey: []
      summary: Get one domain's canonical activity timeline
      tags:
        - Exploration
  /api/v1/explore:
    post:
      operationId: explore
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExploreHTTPRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExploreHTTPResponse"
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Error
        "409":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Error
        "503":
          content:
            application/json:
              schema:
                anyOf:
                  - $ref: "#/components/schemas/ExploreCacheUnavailableResponse"
                  - $ref: "#/components/schemas/ErrorResponse"
          description: Service Unavailable
        default:
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Error
      security:
        - apiKey: []
      summary: Explore canonical archive entries
      tags:
        - Exploration
...
  /api/v1/files/search:
    post:
      operationId: searchFiles
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FileSearchHTTPRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileSearchHTTPResponse"
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Error
        "409":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Error
        "503":
          content:
            application/json:
              schema:
                anyOf:
                  - $ref: "#/components/schemas/ExploreCacheUnavailableResponse"
                  - $ref: "#/components/schemas/ErrorResponse"
          description: Service Unavailable
        default:
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Error
      security:
        - apiKey: []
      summary: Search analytical files
      tags:
        - Exploration
...
  /api/v1/people/search:
    post:
      operationId: searchPeople
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IdentitySearchHTTPRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PersonSearchHTTPResponse"
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Error
        "409":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Error
        "503":
          content:
            application/json:
              schema:
                anyOf:
                  - $ref: "#/components/schemas/ExploreCacheUnavailableResponse"
                  - $ref: "#/components/schemas/ErrorResponse"
          description: Service Unavailable
        default:
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Error
      security:
        - apiKey: []
      summary: Search analytical people
      tags:
        - Exploration
  /api/v1/people/{id}:
    get:
      operationId: getPerson
      parameters:
        - description: Durable participant ID
          in: path
          name: id
          required: true
          schema:
            format: int64
            type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PersonSummary"
          description: OK
        default:
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Error
      security:
        - apiKey: []
      summary: Get one analytical person
      tags:
        - API
  /api/v1/people/{id}/files/search:
    post:
      operationId: searchPersonFiles
      parameters:
        - description: Durable participant ID
          in: path
          name: id
          required: true
          schema:
            format: int64
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FileSearchHTTPRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileSearchHTTPResponse"
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Error
        "409":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Error
        "503":
          content:
            application/json:
              schema:
                anyOf:
                  - $ref: "#/components/schemas/ExploreCacheUnavailableResponse"
                  - $ref: "#/components/schemas/ErrorResponse"
          description: Service Unavailable
        default:
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Error
      security:
        - apiKey: []
      summary: Search one person's analytical files
      tags:
        - Exploration
  /api/v1/people/{id}/summary:
    post:
      operationId: getPersonContextSummary
      parameters:
        - description: Durable participant ID
          in: path
          name: id
          required: true
          schema:
            format: int64
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExploreHTTPRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PersonContextSummaryHTTPResponse"
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Error
        "409":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Error
        "503":
          content:
            application/json:
              schema:
                anyOf:
                  - $ref: "#/components/schemas/ExploreCacheUnavailableResponse"
                  - $ref: "#/components/schemas/ErrorResponse"
          description: Service Unavailable
        default:
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Error
      security:
        - apiKey: []
      summary: Get one person's contextual analytical summary
      tags:
        - Exploration
...